2.0.0

+++ src/main/java/com/proyectofinal_cfp18/choferMapping.java
@@ -2,8 +2,8 @@ package com.proyectofinal_cfp18;
 
 public class choferMapping {
 
-    public chofer mapChofer(String nombreBD, String apellidoBD, long dniDB, int antiguedadDB, long sueldoBaseBD){
-    chofer chofer = new chofer(nombreBD,  apellidoBD,dniDB, antiguedadDB, sueldoBaseBD);
+    public chofer mapChofer(String nombreBD, String apellidoBD, Integer dniDB, Integer antiguedadDB, Integer sueldoBaseBD){
+    chofer chofer = new chofer(nombreBD, apellidoBD,dniDB, antiguedadDB, sueldoBaseBD);
     return chofer;
     
     }
